# openAPI.yml
openapi: "3.0.2"
info:
  title: Сервис друзей
  version: 0.0.1
servers:
  - url: http://127.0.0.1:8000
    description: dev server
paths:
...
  /user/{user_id}/friend/{friend_id}:
    delete:
      summary: Метод удаления друга
      tags:
        - Friends
      operationId: deleteFriend
      parameters:
        - name: user_id
          in: path
          required: true
          description: ID пользователя, который хочет удалить друга
          schema:
            type: integer
          example: 1
        - name: friend_id
          in: path
          required: true
          description: ID друга
          schema:
            type: integer
          example: 2
      responses:
        "200":
          description: Успешный ответ с пользователем, которого удалили
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "default":
          description: Все остальное
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    User:
      type: object
      required:
        - username
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: kirin
...
    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          example: not found
...